+++ tests/integration/scripting_test.go
@@ -46,8 +46,7 @@ func TestScripting(t *testing.T) {
 		ydb.WithLogger(
 			trace.MatchDetails(`ydb\.(driver|discovery|retry|scheme).*`),
 			ydb.WithNamespace("ydb"),
-			ydb.WithOutWriter(logger),
-			ydb.WithErrWriter(logger),
+			ydb.WithWriter(logger),
 			ydb.WithMinLevel(log.TRACE),
 		),
 		ydb.WithUserAgent("scripting"),