+++ tests/integration/table_bulk_upsert_test.go
@@ -127,8 +127,7 @@ func TestTableBulkUpsert(t *testing.T) {
 		ydb.WithLogger(
 			trace.MatchDetails(`ydb\.(driver|discovery|retry|scheme).*`),
 			ydb.WithNamespace("ydb"),
-			ydb.WithOutWriter(logger),
-			ydb.WithErrWriter(logger),
+			ydb.WithWriter(logger),
 			ydb.WithMinLevel(log.TRACE),
 		),
 	)