+++ tests/integration/table_multiple_result_sets_test.go
@@ -51,8 +51,7 @@ func TestTableMultipleResultSets(t *testing.T) {
 		ydb.WithLogger(
 			trace.MatchDetails(`ydb\.(driver|discovery|retry|scheme).*`),
 			ydb.WithNamespace("ydb"),
-			ydb.WithOutWriter(logger),
-			ydb.WithErrWriter(logger),
+			ydb.WithWriter(logger),
 			ydb.WithMinLevel(log.TRACE),
 		),
 	)